dom/xhr/tests/test_bug1300552.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1300552</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script type="application/javascript">

var w = new Worker('worker_bug1300552.js');
w.onmessage = function(e) {
  if (e.data.type == 'info') {
    info(e.data.msg);
  } else if (e.data.type == 'check') {
    ok(e.data.what, e.data.msg);
  } else if (e.data.type == 'finish') {
    SimpleTest.finish();
  } else {
    ok(false, 'Something wrong happened');
  }
}

SimpleTest.waitForExplicitFinish();

  </script>
</body>
</html>
